+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
-1.0.7 - 
+1.1.0 - 
 ------------------
+
+* since 1.1.0 the ``Table`` object is no longer a subclass of
+  list. Records are instead available in the ``records`` attribute,
+  but the table can be iterated over like before. This change was made
+  to make the API cleaner and easier to understand. ``read()`` is
+  still included for backwards compatability, and returns an
+  ``OldStyleTable`` object with the old behaviour.
 
 * Character encoding will now be "latin1" when
   ``language_driver == 0``. This prevents decoding errors when